Detect special version 1.12
There exists a special version of Gothic 1, refered to as 1.12 that
sits somewhere in betwee Gothic 1 and 2.

The datafiles contain specific music files that make it easy to
detect it.

So set the patchlevel to 12 for this version and 0 otherwise.

This version also uses a files named Game.dat instead of Gothic.dat
so handle this as well.