Handle different ai_printscreen variants
ai_printscreen was introduced with Gothic 2 and is not used in vanilla
Gothic 1. Version 1.12 though introduced a different variant with a
different argument list.

So add a variant for it and register the correct one.

Registering the main ai_printscreen also for Gothic 1 does not really
matter, as it's not used there.