Implement ai_quicklook
The quicklook function is meant to have the Npc just move its head to the
player for two seconds.

Testcase in vanilla Gothic 1 was at the exchange place, while Diego walks
back to the old camp. Stand next to where he is walking and pick up one of
the berries, arrows, etc. Diego will continue walking but turn his head
over for those 2 seconds.

So the implementation does the same, not changing the currentLookAtNpc
but storing the temporary one in a separate variable for the duration
of the action. That way this cinematic should not influence any other
parts or save/load mechanisms.